requirements.txt
discord.py
tzdata
yt-dlp
PyNaCl
python-dotenv
bgutil-ytdlp-pot-provider
fastapi
uvicorn[standard]
jinja2
python-multipart
davey
sqlalchemy[asyncio]
aiosqlite